This implements the Caspar CLI
Instead of modules (in TBG-world), Caspar handles things in
terms of namespaces. A cli directory should exist in the module's
directory, if that module is to have CLI commands. The namespace
of CLI command classes in that directory should be \module\cli,
and then cli commands are accessed in terms of module:command.

This does mean modules are used, but we access modules in terms of
the namespace, instead of using a modules table. It is anticipated
that module authors will use other namespaces, i.e. \module\foo
for their own work.

Everything is pretty much automatic now - there is no need to
explicitly state that CLI commands are available. Simply creating
the 'cli' directory is sufficient to get the CLI looking for
commands, any files of the form Grub.class.php, containing a class
\module\cli\Grub inheriting from \caspar\core\CliCommand will be
picked up and be automatically usable.

Core CLI commands go in the caspar\core\cli directory and the class
shall be \caspar\cli\Grub. In both cases, Grub is some unique
identifier within that namespace (and directory) for the command
but is not necessarily the command name, that can be specified
within the command.

Not implemented: authentication for CLI commands (should
authenticate automatically from the POSIX data, needs modifications
within the User object)